+++ location/ios/Classes/LocationPlugin.m
@@ -15,6 +15,7 @@
 @property (copy, nonatomic)   FlutterEventSink   flutterEventSink;
 @property (assign, nonatomic) BOOL               flutterListening;
 @property (assign, nonatomic) BOOL               hasInit;
+@property (assign, nonatomic) BOOL               applicationHasLocationBackgroundMode;
 @end
 
 @implementation LocationPlugin
@@ -48,6 +49,9 @@
     if (!(self.hasInit)) {
         self.hasInit = YES;
 
+        NSArray *backgroundModes = [NSBundle.mainBundle objectForInfoDictionaryKey:@"UIBackgroundModes"];
+        self.applicationHasLocationBackgroundMode = [backgroundModes containsObject: @"location"];
+
         if ([CLLocationManager locationServicesEnabled]) {
             self.clLocationManager = [[CLLocationManager alloc] init];
             self.clLocationManager.delegate = self;
@@ -76,6 +80,21 @@
             }
             self.clLocationManager.distanceFilter = distanceFilter;
             result(@1);
+        }
+    } else if ([call.method isEqualToString:@"isBackgroundModeEnabled"]) {
+        if (self.applicationHasLocationBackgroundMode) {
+            result(self.clLocationManager.allowsBackgroundLocationUpdates ? @1 : @0);
+        } else {
+            result(@0);
+        }
+    } else if ([call.method isEqualToString:@"enableBackgroundMode"]) {
+        BOOL enable = [call.arguments[@"enable"] boolValue];
+        if (self.applicationHasLocationBackgroundMode) {
+            self.clLocationManager.allowsBackgroundLocationUpdates = enable;
+            self.clLocationManager.showsBackgroundLocationIndicator = enable;
+            result(enable ? @1 : @0);
+        } else {
+            result(@0);
         }
     } else if ([call.method isEqualToString:@"getLocation"]) {
         if (![CLLocationManager locationServicesEnabled]) {