+++ location_platform_interface/CHANGELOG.md
@@ -1,3 +1,6 @@
+## [1.1.0] - 07th December 2020
+- Add new interface method enableBackgroundMode(boolean) to control weather plugin should work in the background or not.
+
 ## [1.0.1] - 24th August 2020
 
 - Add grantedLimited enum value to PermissionStatus to capture the new limited range authorization that iOS 14 is offering.